lib: move functions to mbyte.c
